+++ CHANGELOG.md
@@ -47,3 +47,4 @@ the feature for all Things.
 - **FOV effects not being cleared thoroughly upon loading levels**
 - **Turbo cheat setting incorrect walk-strafing speed**
 - **Crash when toggling Alt. Arms Display without entering a map**
+- **"Seizure"-wipe softlock when using it while the screen is darkened**